+++ src/components/status/index.tsx
@@ -13,7 +13,8 @@ export default function Status({
     <div className=" text-xs font-medium whitespace-nowrap">
       {status?.toLocaleLowerCase().includes("not") ||
       status?.toLocaleLowerCase().includes("cancel") ||
-      status?.toLocaleLowerCase().includes("closed") ? (
+      status?.toLocaleLowerCase().includes("closed") ||
+      status?.toLocaleLowerCase().includes("decline") ? (
         <span className=" p-1 px-3 bg-red-500/15 text-red-500 rounded-full">
           {status}
         </span>