+++ src/components/tables/requisition-request-table.tsx
@@ -8,17 +8,19 @@ import ExportSelect from "../inputs/select/exportSelect";
 import Select from "../inputs/select";
 import useGetRequisitionRequests from "../../services-hooks/useGetRequisitionRequests";
 import Status from "../status";
-import DoubleCheckIcon from "../../assets/icons/double-check";
 import PenIcon from "../../assets/icons/pen";
 import LoadingButton from "../button";
 import PlusIcon from "../../assets/icons/plus";
 import ModalTemplate from "../modal";
 import AddEditRequisitionRequest from "../../pages/requests/requisition-requests/add-edit-requisition-requests";
+import RequisitionRequestStatusChanger from "../../pages/requests/requisition-requests/staus-update";
+import DoubleCheckIcon from "../../assets/icons/double-check";
+import { requisitionRequest } from "../../types/apiData/requisition-request";
 // import BinIcon from "../../assets/icons/bin-icon";
 // import DeleteConfirmation from "../infoModal/delete-confirmation";
-import useAxios from "../../useHooks/useAxios";
-import { useAppDispatch } from "../../stores/hooks";
-import { removeRequisitionRequestInList } from "../../stores/apiData/requisition-requests";
+// import useAxios from "../../useHooks/useAxios";
+// import { useAppDispatch } from "../../stores/hooks";
+// import { removeRequisitionRequestInList } from "../../stores/apiData/requisition-requests";
 
 export default function RequisitionRequestTable() {
   // const axios = useAxios({ disableErrMssg: false, disableSuccMssg: false });
@@ -32,7 +34,12 @@ export default function RequisitionRequestTable() {
   }>();
   const [currentPage, setCurrentPage] = useState(1);
   const [selectedId, setSelectedId] = useState("");
+  const [requisitionItem, setRequistionItem] = useState<requisitionRequest>(
+    {} as any
+  );
   const [openRequest, setOpenRequest] = useState(false);
+
+  const [openStatusUpdate, setOpenStatusUpdate] = useState(false);
   // const [openDelete, setOpenDelete] = useState(false);
   // const [isDeleting, setIsDeleting] = useState(false);
 
@@ -54,11 +61,15 @@ export default function RequisitionRequestTable() {
     setOpenRequest(true);
   }, []);
 
-  const openForEdit = useCallback((id: number) => {
-    setSelectedId(String(id || ""));
+  const openForEdit = useCallback((item: requisitionRequest) => {
+    setRequistionItem(item);
     setOpenRequest(true);
   }, []);
 
+  const openForStatusUpdate = useCallback((id: number) => {
+    setSelectedId(String(id));
+    setOpenStatusUpdate(true);
+  }, []);
   // const handleOpenDelete = useCallback((id: number) => {
   //   setSelectedId(String(id) || "");
   //   setOpenDelete(true);
@@ -123,6 +134,7 @@ export default function RequisitionRequestTable() {
                       "Amount",
                       "Currency",
                       "Request date",
+                      "Paid",
                       "Status",
                       "Action",
                     ].map((head) => (
@@ -154,16 +166,27 @@ export default function RequisitionRequestTable() {
                         <td>{item?.currency}</td>
                         <td>{formatDate(item?.created_at)}</td>
                         <td>
+                          <Status
+                            status={"requisition_payment_state"}
+                            booleanVal={Boolean(item?.is_paid)}
+                            truthyMessage="Yes"
+                            falsyMessage="No"
+                          />
+                        </td>
+                        <td>
                           <Status status={item?.status} />
                         </td>
                         <td>
                           <div className=" flex items-center gap-4">
-                            <button title="mark as paid">
+                            <button
+                              onClick={() => openForStatusUpdate(item?.id)}
+                              title="mark as paid"
+                            >
                               <DoubleCheckIcon className=" size-8" />
                             </button>
                             <button
                               title="edit"
-                              onClick={() => openForEdit(item?.id)}
+                              onClick={() => openForEdit(item)}
                             >
                               <PenIcon />
                             </button>
@@ -209,7 +232,23 @@ export default function RequisitionRequestTable() {
         title="Requisition"
         className=" max-w-screen-md"
       >
-        <AddEditRequisitionRequest id={selectedId} setOpen={setOpenRequest} />
+        <AddEditRequisitionRequest
+          requisitionItem={requisitionItem}
+          id={String(requisitionItem?.id || "")}
+          setOpen={setOpenRequest}
+        />
+      </ModalTemplate>
+      <ModalTemplate
+        open={openStatusUpdate}
+        setOpen={setOpenStatusUpdate}
+        showXicon={true}
+        title="Requisition Status Update"
+        className=" max-w-screen-md"
+      >
+        <RequisitionRequestStatusChanger
+          id={selectedId}
+          setOpen={setOpenStatusUpdate}
+        />
       </ModalTemplate>
     </>
   );