+++ src/stores/apiData/requisition-requests.ts
@@ -102,6 +102,37 @@ export const requisitionRequestsListData = createSlice({
       });
       state.value.pagination = replacedItem;
     },
+    updateRequisitionRequestStatusInList: (state, action) => {
+      const { id } = action?.payload;
+      const currentArray = state.value.data;
+      const currentIndex = currentArray.findIndex(
+        (v: { id: number }) => String(v.id) === String(id)
+      );
+      const currentDataset = currentArray[currentIndex];
+      if (currentIndex >= 0) {
+        currentArray.splice(currentIndex, 1, {
+          ...currentDataset,
+          ...action?.payload,
+        });
+        state.value.data = currentArray;
+      }
+      //REPLACE pagination data
+      const pagination_data = [...state.value.pagination];
+      const replacedItem = pagination_data.map((item, index) => {
+        const sencondFilter = item.data.map((data, i) => {
+          if (String(data.id) === String(id)) {
+            return { ...currentDataset, ...action?.payload };
+          } else {
+            return data;
+          }
+        });
+        return {
+          pagination_data: { ...item.pagination_data },
+          data: sencondFilter,
+        };
+      });
+      state.value.pagination = replacedItem;
+    },
     clearRequisitionRequestList: (state) => {
       state.value.status = false;
       state.value.data = [];
@@ -111,6 +142,7 @@ export const requisitionRequestsListData = createSlice({
 
 export const {
   updateRequisitionRequestList,
+  updateRequisitionRequestStatusInList,
   addRequisitionRequestToList,
   addToPaginationHistory,
   removeRequisitionRequestInList,